[ refactor ] core 빼기
