+++ src/components/CardCollection/FilterModal/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
 
+import { filterTags, intimacyTags } from "../../../core/cardCollection/filter";
 import Modal from "../../common/Modal";
 import IntimacySlider from "./IntimacySlider";
 import { St } from "./style";
@@ -8,38 +9,18 @@ interface FilterModalProps {
   closeHandler: () => void;
 }
 
-type FilterTags = {
-  type: string;
-  tags: string[];
-};
-
-const filterTags: FilterTags[] = [
-  {
-    type: "성별",
-    tags: ["남", "여"],
-  },
-  {
-    type: "연령대",
-    tags: ["10대", "20대", "30대"],
-  },
-  {
-    type: "술자리 유형",
-    tags: ["개인", "커플", "친구", "단체"],
-  },
-];
-
-const intimacyTags: string[] = ["상관없음", "새로워요", "친근해요", "절친해요"];
-
 export default function FilterModal(props: FilterModalProps) {
   const { closeHandler } = props;
   const [checkedTags, setCheckedTags] = useState<Set<string>>(new Set()); // 체크한 태그들을 저장할 state
   const [intimacyValues, setIntimacyValues] = useState<number[]>([0]); // 친밀도 value
+
   // 태그를 눌렀을 때 함수
   const toggleTag = (_tag: string) => {
     const tempCheckedTags = new Set([...checkedTags]);
     tempCheckedTags.has(_tag) ? tempCheckedTags.delete(_tag) : tempCheckedTags.add(_tag);
     setCheckedTags(tempCheckedTags);
   };
+
   // 추천 시작하기를 눌렀을 때, 태그 정보들과 친밀도 정보를 담아주고 창닫기
   const submitFilter = () => {
     const tempCheckedTags = new Set(checkedTags);
