+++ src/core/cardCollection/filter.ts
@@ -1,0 +1,21 @@
+type FilterTags = {
+  type: string;
+  tags: string[];
+};
+
+export const filterTags: FilterTags[] = [
+  {
+    type: "성별",
+    tags: ["남", "여"],
+  },
+  {
+    type: "연령대",
+    tags: ["10대", "20대", "30대"],
+  },
+  {
+    type: "술자리 유형",
+    tags: ["개인", "커플", "친구", "단체"],
+  },
+];
+
+export const intimacyTags: string[] = ["상관없음", "새로워요", "친근해요", "절친해요"];
